feat: add validations for booking

+++ src/main/java/com/flights/service/BookingServiceImpl.java
@@ -5,6 +5,7 @@ import com.flights.bean.Passenger;
 import com.flights.bean.ScheduledFlight;
 import com.flights.dao.BookingDao;
 import com.flights.exception.InvalidPassengerUIN;
+import com.flights.exception.RecordNotFound;
 import com.flights.exception.SeatNotAvailable;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -23,15 +24,19 @@ public class BookingServiceImpl implements  BookingService{
     ScheduledFlight scheduledFlightRepo;
 
     @Override
-    public Booking addBooking(Booking booking) {
-        // validateBooking(booking);
+    public Booking addBooking(Booking booking) throws Exception {
+        validateBooking(booking);
+        for (Passenger p: booking.getPassengerList())
+            validatePassenger(p);
         repo.save(booking);
         return booking;
     }
 
     @Override
-    public Booking modifyBooking(Booking booking) {
+    public Booking modifyBooking(Booking booking) throws RecordNotFound {
         int id = booking.getBookingId();
+        if(repo.findById(id) == null)
+            throw new RecordNotFound("Booking object does not exist");
         Booking b = repo.findById(id).orElseThrow();
         b.setBookingDate(booking.getBookingDate());
         b.setPassengerList(booking.getPassengerList());
@@ -41,7 +46,9 @@ public class BookingServiceImpl implements  BookingService{
     }
 
     @Override
-    public Booking viewBooking(int bookingId) {
+    public Booking viewBooking(int bookingId) throws RecordNotFound {
+        if(repo.findById(bookingId) == null)
+            throw new RecordNotFound("Booking object does not exist");
         Booking b = repo.findById(bookingId).orElseThrow();
         return b;
     }
@@ -53,7 +60,9 @@ public class BookingServiceImpl implements  BookingService{
     }
 
     @Override
-    public void deleteBooking(int bookingId) {
+    public void deleteBooking(int bookingId) throws RecordNotFound {
+        if(repo.findById(bookingId) == null)
+            throw new RecordNotFound("Booking object does not exist");
         Booking b = repo.findById(bookingId).orElseThrow();
         repo.delete(b);
     }
